% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{addf}
\alias{addf}
\title{Create xts addfactor}
\usage{
addf(
  start = bnk_start,
  end = bnk_end,
  from = 0,
  to = 0,
  ser_name = "value",
  per = "year"
)
}
\arguments{
\item{start}{start date of linear interpolation (character: "yyyy-mm-dd", "yyyyqq", "yyyy")}

\item{end}{end date of linear interpolation (character: "yyyy-mm-dd", "yyyyqq", "yyyy")}

\item{from}{first value for linear interpolation (numeric)}

\item{to}{last value for linear interpolation (numeric)}

\item{ser_name}{name of the xts series (string)}

\item{per}{periodicity of series (character: "year" - default)
if date format of start is quarterly, automatically set to "quarter"}
}
\value{
a single xts series spanning bnk_start-bnk_end
}
\description{
Create xts addfactor
}
\details{
this is a wrapper around make_xts with some additional functionality.
the start and end dates specify the span of the non-zero add-factor value. the
remaining period between start and end is filled with zeros.
}
\examples{
addf()
addf(201002, 201504, 1, 2)
addf(20100101, 20601201, 1, 2, per = "month")
addf(20100101, from = 1, to = 2, per = "quarter")
addf(2010.2, 2015.4, 1, 2, "ECT_HI")
}
