% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{aggr}
\alias{aggr}
\title{Aggregate univariate or multivariate time series from low to high frequency}
\usage{
aggr(x, conv_type = "mean", target_freq = "year", na_rm = FALSE)
}
\arguments{
\item{x}{a tx-boxable object at a high frequency (e.g. monthly or quarterly)}

\item{conv_type}{match the aggregated value via "first", "last", "sum",
"mean". If conv_type == "uhero" then the name of the time series x is
compared to the internal variable \code{sum_pattern}. For matching series names
the aggregation is based on "sum"; for all others it is based on "mean."}

\item{target_freq}{target frequency "year", "quarter", "month", "week"}

\item{na_rm}{logical, if TRUE, incomplete periods are aggregated as
well. For irregular series, incomplete periods are always aggregated.}
}
\value{
aggregated object of the same type as the input
}
\description{
Aggregate univariate or multivariate time series from low to high frequency
}
\examples{
monthly_data_example |>
  aggr(conv_type = "sum", target_freq = "quarter")
monthly_data_example |>
  aggr(conv_type = "uhero", target_freq = "quarter")
monthly_data_example |>
  aggr(conv_type = "uhero", target_freq = "quarter") |>
  tsbox::ts_long() |>
  disagg(conv_type = "uhero", target_freq = "month") |>
  tsbox::ts_wide() # this is close to original data
# works with a single series too
monthly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_pick("VISNS_HI") |>
  aggr(conv_type = "uhero", target_freq = "year") |>
  tsbox::ts_plot()
}
