% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{AtoQ}
\alias{AtoQ}
\title{Linear interpolation based on AREMOS command reference page 292
(superseded by disagg())}
\usage{
AtoQ(ser_in, aggr = "mean")
}
\arguments{
\item{ser_in}{the xts series to be interpolated (freq = a)}

\item{aggr}{interpolation method: aggregate via mean (default) or sum}
}
\value{
interpolated xts series (freq = q)
}
\description{
Linear interpolation based on AREMOS command reference page 292
(superseded by disagg())
}
\examples{
quarterly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_xts() |>
  tsbox::ts_pick("E_NF_HI") |>
  QtoA() |> # this matches with below
  AtoQ() |>
  QtoA() |> # this matches with above
  tsbox::ts_plot()
}
