% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data.R
\docType{data}
\name{bnk_end}
\alias{bnk_end}
\title{end of data range in the data bank}
\format{
scalar yyyy-mm-dd date
}
\source{
author
}
\usage{
bnk_end
}
\description{
end of data range in the data bank
}
\references{
\url{uhero.hawaii.edu}
}
\author{
Peter Fuleky \email{fuleky@hawaii.edu}
}
\keyword{data}
