% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data.R
\docType{data}
\name{bnk_start}
\alias{bnk_start}
\title{define project-wide constants}
\format{
scalar yyyy-mm-dd date
}
\source{
author
}
\usage{
bnk_start
}
\description{
start of data range in the data bank
}
\references{
\url{uhero.hawaii.edu}
}
\author{
Peter Fuleky \email{fuleky@hawaii.edu}
}
\keyword{data}
