% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expr_data.R
\name{clean_file_names}
\alias{clean_file_names}
\title{Clean file names by removing longest common substring from filename}
\usage{
clean_file_names(x)
}
\arguments{
\item{x}{names to be cleaned}
}
\value{
string of same dimension of x, with longest common substring removed
}
\description{
Clean file names by removing longest common substring from filename
}
