% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_from_middle.R
\name{color_from_middle}
\alias{color_from_middle}
\title{modify coloring in DT::datatable json
based on work by \code{NicE}, posted here on SO: https://stackoverflow.com/a/33524422/3457743}
\usage{
color_from_middle(data, color1, color2)
}
\description{
modify coloring in DT::datatable json
based on work by \code{NicE}, posted here on SO: https://stackoverflow.com/a/33524422/3457743
}
