% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_utils.R
\name{common_substrings}
\alias{common_substrings}
\title{Find all common substrings of two strings
Based on https://stackoverflow.com/questions/28261825/longest-common-substring-in-r-finding-non-contiguous-matches-between-the-two-str}
\usage{
common_substrings(a, b)
}
\arguments{
\item{a}{first string}

\item{b}{second string}
}
\description{
Find all common substrings of two strings
Based on https://stackoverflow.com/questions/28261825/longest-common-substring-in-r-finding-non-contiguous-matches-between-the-two-str
}
