% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{conv_long}
\alias{conv_long}
\title{Convert "ts-boxable" objects into long format}
\usage{
conv_long(x, ser_info = FALSE)
}
\arguments{
\item{x}{a "tx-boxable" object to be converted}

\item{ser_info}{should additional details be returned (TRUE) or
only the long format of x (default: FALSE)}
}
\value{
returns a ts-boxable object in long format with \code{id}, \code{time} and
\code{value} columns. if \code{ser_info = TRUE}, also returns the following attributes:
\code{was_wide} is \code{TRUE} if x is a wide data frame, \code{FALSE} otherwise, and
\code{ser_names} are the names of the series in x.
}
\description{
Convert "ts-boxable" objects into long format
}
\details{
This function converts wide data frames and other ts-boxable objects
to the long format (wide data frames are not ts-boxable). In addition, it
ensures that objects containing a single time series have an id column.
}
\examples{
quarterly_data_example |>
  conv_long()
quarterly_data_example |>
  conv_long() |>
  tsbox::ts_tslist() |>
  conv_long()
quarterly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_xts() |>
  conv_long(ser_info = TRUE)
quarterly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_pick("E_NF_HI") |>
  tsbox::ts_xts() |>
  conv_long()
quarterly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_xts() |>
  tsbox::ts_pick("E_NF_HI") |>
  conv_long()
}
