% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{conv_tslist}
\alias{conv_tslist}
\title{Convert "ts-boxable" objects to tslist (extension of \code{tsbox::ts_tslist()})}
\usage{
conv_tslist(x)
}
\arguments{
\item{x}{a "ts-boxable" object to be converted}
}
\value{
returns an object as a tslist.
}
\description{
Convert "ts-boxable" objects to tslist (extension of \code{tsbox::ts_tslist()})
}
\details{
This function performs a similar operation to \code{tsbox::ts_tslist()}.
It converts ts-boxable objects to the tslist format. An important difference
compared with \code{tsbox::ts_tslist()} is that the x argument of \code{conv_tslist()}
can be a wide tbl.
}
\examples{
quarterly_data_example |>
  conv_xts() |>
  conv_tslist()
quarterly_data_example |>
  conv_wide() |>
  conv_tslist()
quarterly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_pick("E_NF_HI") |>
  conv_wide() |>
  conv_tslist()
}
