% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{conv_xts}
\alias{conv_xts}
\title{Convert "ts-boxable" objects to xts format (extension of \code{tsbox::ts_xts()})}
\usage{
conv_xts(x)
}
\arguments{
\item{x}{a "ts-boxable" object to be converted}
}
\value{
returns an object in xts format.
}
\description{
Convert "ts-boxable" objects to xts format (extension of \code{tsbox::ts_xts()})
}
\details{
This function performs a similar operation to \code{tsbox::ts_xts()}. It
converts ts-boxable objects to the xts format. An important difference
compared with \code{tsbox::ts_xts()} is that the x argument of \code{conv_xts()}
can be a wide tbl.
}
\examples{
quarterly_data_example |>
  conv_tslist() |>
  conv_xts()
quarterly_data_example |>
  conv_wide() |>
  conv_xts()
quarterly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_pick("E_NF_HI") |>
  conv_wide() |>
  conv_xts()
}
