% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expr_data.R
\name{create_deseq2_result}
\alias{create_deseq2_result}
\title{summarize expr data as DESeqDataSet
importFrom DESeq2 DESeqDataSetFromMatrix}
\usage{
create_deseq2_result(fc, sample_data, design = ~cond)
}
\description{
summarize expr data as DESeqDataSet
importFrom DESeq2 DESeqDataSetFromMatrix
}
