% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expr_data.R
\name{create_edger_result}
\alias{create_edger_result}
\title{Summarise expr data as DGEList object
importFrom edgeR DGEList}
\usage{
create_edger_result(fc, sample_data)
}
\description{
Summarise expr data as DGEList object
importFrom edgeR DGEList
}
