% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data.R
\docType{data}
\name{daily_data_example}
\alias{daily_data_example}
\title{daily data for examples}
\format{
tibble of daily data
}
\source{
author
}
\usage{
daily_data_example
}
\description{
daily data for examples
}
\references{
\url{uhero.hawaii.edu}
}
\author{
Peter Fuleky \email{fuleky@hawaii.edu}
}
\keyword{data}
