% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{disagg}
\alias{disagg}
\title{Interpolate univariate or multivariate time series from low to high frequency}
\usage{
disagg(x, conv_type = "mean", target_freq = "quarter")
}
\arguments{
\item{x}{a tx-boxable object at a low frequency (e.g. annual or quarterly)}

\item{conv_type}{match the quarterly value via "first", "last", "sum", "mean"}

\item{target_freq}{target frequency "quarter" or "month"}
}
\value{
interpolated object of the same type as the input
}
\description{
Interpolate univariate or multivariate time series from low to high frequency
}
\examples{
quarterly_data_example |>
  disagg(conv_type = "mean", target_freq = "month")
quarterly_data_example |>
  disagg(conv_type = "mean", target_freq = "month") |>
  tsbox::ts_long() |>
  tsbox::ts_frequency(to = "quarter", aggregate = "mean") |>
  tsbox::ts_wide() # this matches original data
quarterly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_pick("E_NF_HI") |>
  disagg(conv_type = "mean", target_freq = "month") |>
  tsbox::ts_plot() # works with a single series too
}
