% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{faggr}
\alias{faggr}
\title{Aggregate univariate or multivariate fiscal time series from low to high frequency}
\usage{
faggr(x, conv_type = "mean", target_freq = "year", na_rm = FALSE)
}
\arguments{
\item{x}{a tx-boxable object at a high frequency (e.g. monthly or quarterly)}

\item{conv_type}{match the aggregated value via "first", "last", "sum",
"mean". If conv_type == "uhero" then the name of the time series x is
compared to the internal variable \code{sum_pattern}. For matching series names
the aggregation is based on "sum"; for all others it is based on "mean."}

\item{target_freq}{target frequency "year", "quarter", "month", "week"}

\item{na_rm}{logical, if TRUE, incomplete periods are aggregated as
well. For irregular series, incomplete periods are always aggregated.}
}
\value{
aggregated object of the same type as the input
}
\description{
Aggregate univariate or multivariate fiscal time series from low to high frequency
}
\details{
date stamps typically reflect the start of a period, but here the
date stamps in the result are shifted 6 months forward relative to calendar
dates, so Jul21 becomes Jan22, or 2022-01-01, and 21Q3 becomes 22Q1.
}
\examples{
monthly_data_example |>
  faggr(conv_type = "sum", target_freq = "quarter")
monthly_data_example |>
  faggr(conv_type = "uhero", target_freq = "quarter")
# works with a single series too
monthly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_pick("VISNS_HI") |>
  faggr(conv_type = "uhero", target_freq = "year") |>
  tsbox::ts_plot()
}
