% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{fcutils_options}
\alias{fcutils_options}
\title{Set or get options for fcutils package}
\usage{
fcutils_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for fcutils package
}
\section{Supported options}{

The following options are supported
\itemize{
\item{\code{cache_dir}}{(\code{character};.cache_dir) Path used as local file cache}
}
}

