% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{fcutils_options_reset}
\alias{fcutils_options_reset}
\title{Reset global options for fcutils}
\format{An object of class \code{list} of length 1.}
\usage{
fcutils_options_reset
}
\description{
Reset global options for fcutils
}
\keyword{datasets}
