% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_counts.R
\name{filter_low_counts}
\alias{filter_low_counts}
\title{Filter out genes/transcripts with low expression}
\usage{
filter_low_counts(e, cpm = TRUE, min_value = 0)
}
\arguments{
\item{e}{expression data}

\item{cpm}{filter on cpm threshold default, otherwise use min_value}

\item{min_value}{if cpm is false, remove transcripts with rowsum <= this}
}
\description{
Filter out genes/transcripts with low expression
}
