% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{find_start}
\alias{find_start}
\title{Find the date of the first observation (NAs are dropped)}
\usage{
find_start(ser_in)
}
\arguments{
\item{ser_in}{an xts series}
}
\value{
date associated with first observation
}
\description{
Find the date of the first observation (NAs are dropped)
}
\examples{
`ncen@us.sola` <- ts(NA_real_, start = 2016, end = 2021, freq = 1) |> tsbox::ts_xts()
`ncen@us.sola`["2017/2021"] <- c(325511184, 327891911, 330268840, 332639102, 334998398)
find_start(`ncen@us.sola`)
}
