% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{fytd_cum}
\alias{fytd_cum}
\title{Fiscal year to date sum or average}
\usage{
fytd_cum(x, avg = TRUE)
}
\arguments{
\item{x}{a ts-boxable object}

\item{avg}{if TRUE (default), return year to date average, if FALSE, return year to date sum}
}
\value{
object of the same type as the input containing year to date sum or average
}
\description{
Fiscal year to date sum or average
}
\details{
this function operates similarly to ytd_cum() but assumes that the
fiscal year starts in July and accumulates the values from that month onward.
}
\examples{
monthly_data_example |>
  fytd_cum()
quarterly_data_example |>
  fytd_cum()
monthly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_pick("VISNS_HI") |>
  tsbox::ts_xts() |>
  fytd_cum(avg = FALSE) |>
  tsbox::ts_plot()
}
