% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{fytd_gr}
\alias{fytd_gr}
\title{Fiscal year to date growth rate}
\usage{
fytd_gr(x)
}
\arguments{
\item{x}{a ts-boxable object}
}
\value{
object of the same type as the input containing year to date growth rate
}
\description{
Fiscal year to date growth rate
}
\details{
this function operates similarly to ytd_gr() but assumes that the
fiscal year starts in July and accumulates the values from that month onward.
}
\examples{
quarterly_data_example |>
  fytd_gr()
monthly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_pick("VISNS_HI") |>
  tsbox::ts_xts() |>
  fytd_gr() |>
  tail()
}
