% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expr_data.R
\name{get_expr_data}
\alias{get_expr_data}
\title{Summarise expression data as DGEList object}
\usage{
get_expr_data(using = c("DESeq2", "edgeR"), sample_data = NULL,
  design = NULL)
}
\description{
Summarise expression data as DGEList object
}
