% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expr_data.R
\name{get_featurecounts_files}
\alias{get_featurecounts_files}
\title{List files containing featurecounts summaries by gene}
\usage{
get_featurecounts_files(featurecounts_dir, pattern = NULL,
  expect_length = NULL, sample_name_from_filename = basename)
}
\arguments{
\item{featurecounts_dir}{path to directory containing featureCounts output files}

\item{pattern}{If provided, regex pattern for results files (e.g.: 'geneID$')}

\item{expect_length}{If provided, code will confirm the number of result files identified}

\item{sample_name_from_filename}{function to derive sample name from file name. Defaults to \code{basename}, but a useful alternative might be \code{dirname}.}
}
\value{
named list of filepaths in directory matching regex, where name is extracted from the cleaned filename
}
\description{
List files containing featurecounts summaries by gene
}
