% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expr_data.R
\name{get_gene_names}
\alias{get_gene_names}
\title{get gene names from entrez or ensembl ids}
\usage{
get_gene_names(ids, column = "SYMBOL", keytype = c("ENSEMBL",
  "ENTREZID"), host = fcutils_options()$ensembl_host, ...)
}
\description{
get gene names from entrez or ensembl ids
}
