% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{get_series_1}
\alias{get_series_1}
\title{Download a single series from udaman using series name}
\usage{
get_series_1(ser_id, expand = "true", rename = "compact", descr = FALSE)
}
\arguments{
\item{ser_id}{udaman series name}

\item{expand}{"true" or "raw" ("true" downloads formatted data, "raw" downloads raw units)}

\item{rename}{"compact" (default), "full", "no". "compact": @ replaced by _ and no frequency;
"full": @ replaced by __ and . by _;
"no": no renaming, keep UDAMAN names}

\item{descr}{if TRUE add to the udaman series name the series description in parentheses}
}
\value{
time and data for a single series combined in a tibble
}
\description{
Download a single series from udaman using series name
}
\examples{
\dontrun{
get_series_1(ser_id = "VISNS@HI.M")
}
}
