% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{get_series_exp}
\alias{get_series_exp}
\title{Download series listed in an export table from udaman}
\usage{
get_series_exp(
  exp_id,
  format = "wide",
  expand = "true",
  raw = FALSE,
  rename = "compact",
  descr = FALSE,
  public = FALSE,
  save_loc = NULL
)
}
\arguments{
\item{exp_id}{export id (character or numeric)}

\item{format}{"wide" (default) or "long" or "xts"}

\item{expand}{DEPRECATED, USE raw INSTEAD "true" or "raw" ("true" downloads formatted data, "raw" downloads raw units)}

\item{raw}{TRUE or FALSE (default) (TRUE downloads raw data, FALSE downloads scaled and rounded data)}

\item{rename}{"compact" (default), "full", "no". "compact": @ replaced by _
and no frequency; "full": @ replaced by \emph{AT} and . by _; "no": no renaming,
keep UDAMAN names}

\item{descr}{if TRUE add to the udaman series name the series description in parentheses (default: FALSE)}

\item{public}{if TRUE use the public API interface - does not require VPN (default: FALSE)}

\item{save_loc}{file path for saving data incl. extension ("html" or "csv") (default NULL)}
}
\value{
time and data for all series combined in a tibble
}
\description{
Download series listed in an export table from udaman
}
\details{
This function requires permission to access UDAMAN.
Store the udaman token in the .Renviron file using the following format:
udaman_token = "-ABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890="
Or using fcutils::set_udaman_token("-ABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890=")
Or store the udaman token among your credentials (e.g. keychain) using keyring:
keyring::key_set_with_value(service = "udaman_token", password = "-ABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890=")
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_series_exp(exp_id = 74)
get_series_exp(74, format = "xts")
\dontshow{\}) # examplesIf}
}
