% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{\%+=\%}
\alias{\%+=\%}
\title{In place addition}
\usage{
e1 \%+=\% e2
}
\arguments{
\item{e1}{first addend (and returned sum)}

\item{e2}{second addend}
}
\value{
sum of the two addends replacing the values in the first addend
}
\description{
Warning: Typing x \%+=\% y/2 returns x <- (x + y)/2.
Adding parentheses, i.e. x \%+=\% (y/2) solves the problem.
}
\examples{
\dontrun{
add_QMOD.xts$VISUS_HI[pq(2022.3, 2023.4)] <- add_QMOD.xts$VISUS_HI[pq(2022.3, 2023.4)] +
  c(0.01, -0.04, rep(-0.025, 4))
add_QMOD.xts$VISUS_HI[pq(2022.3, 2023.4)] \%+=\% c(0.01, -0.04, rep(-0.025, 4)) # easier on the eye
}
}
