% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{index}
\alias{index}
\title{Get indexed series (wrapper around tsbox::ts_index())}
\usage{
index(x, base_per = as.character(Sys.Date()), base_value = 100)
}
\arguments{
\item{x}{ts-boxable object to be indexed}

\item{base_per}{base date when the index is set to base_value (see examples).
If two dates are provided, the mean in the range is set equal to base_value.}

\item{base_value}{numeric value of the index at base_per (e.g. 1 or 100)}
}
\value{
indexed object of the same type as the input
}
\description{
Get indexed series (wrapper around tsbox::ts_index())
}
\examples{
quarterly_data_example |>
  index(2010.1)
quarterly_data_example |>
  index(c(2010.1, 2010.4))
quarterly_data_example |>
  index(c("2010-01-01", "2010-12-31"), 1)
}
