% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{interpol}
\alias{interpol}
\title{Interpolate univariate or multivariate time series with missing values}
\usage{
interpol(incomplete, pattern, extend_range = FALSE)
}
\arguments{
\item{incomplete}{a ts-boxable object with NAs}

\item{pattern}{a ts-boxable object no missing values}

\item{extend_range}{logical. If TRUE, uses a full_join to potentially extend
the date range. If FALSE (default), uses a left_join to only fill NAs
within the original date range.}
}
\value{
interpolated object of the same type as the input
}
\description{
Interpolate univariate or multivariate time series with missing values
}
\details{
the function aligns two sets of time series by date and then
interpolates the first set using the second set as a pattern.
The time-span of the pattern has to match or be larger
than the time-span of the incomplete data.
}
\examples{
monthly_data_example |>
  dplyr::filter(time > "2002-01-01") |>
  tsbox::ts_long() |>
  dplyr::mutate(
    value = dplyr::if_else(time > "2021-01-01" & time < "2022-01-01",
    NA_real_, value)
  )|>
  interpol(monthly_data_example, extend_range = TRUE)
monthly_data_example |>
  tsbox::ts_long() |>
  dplyr::mutate(
    value = dplyr::if_else(
      time > "2021-01-01" & time < "2022-01-01",
      NA_real_,
      value
    )
  ) |>
  interpol(
    # missing values in the pattern
    monthly_data_example |>
      tsbox::ts_long() |>
      dplyr::mutate(
        value = dplyr::if_else(
          time > "2021-03-01" & time < "2021-09-01",
          NA_real_,
          value
        )
      )
  ) \%>\%
  # no interpolation where pattern has NAs
  tsbox::ts_wide()
}
