% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{interpol_1}
\alias{interpol_1}
\title{Interpolate a single vector containing missing values based on a pattern}
\usage{
interpol_1(vec_incomplete, vec_pattern)
}
\arguments{
\item{vec_incomplete}{numeric vector with NAs to be filled}

\item{vec_pattern}{numeric vector representing the pattern}
}
\value{
numeric vector with NAs filled
}
\description{
Interpolate a single vector containing missing values based on a pattern
}
