% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{interpolate_vector}
\alias{interpolate_vector}
\title{Interpolate a single vector containing missing values based on a pattern}
\usage{
interpolate_vector(vec_incomplete, vec_pattern)
}
\arguments{
\item{vec_incomplete}{numeric vector with NAs to be filled}

\item{vec_pattern}{numeric vector representing the pattern}
}
\value{
numeric vector with NAs filled
}
\description{
Interpolate a single vector containing missing values based on a pattern
}
