% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{is_long}
\alias{is_long}
\title{Check if a data frame is in long format}
\usage{
is_long(x)
}
\arguments{
\item{x}{tibble or data frame}
}
\value{
returns TRUE for long format data frame (id, time, value columns), FALSE otherwise
}
\description{
Check if a data frame is in long format
}
\examples{
monthly_data_example |> is_long()
monthly_data_example |> tsbox::ts_long() |> is_long()
}
