% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{is_wide}
\alias{is_wide}
\title{Check if a data frame is in wide format}
\usage{
is_wide(x)
}
\arguments{
\item{x}{tibble or data frame}
}
\value{
returns TRUE for wide format data frame (time and value columns), FALSE otherwise
}
\description{
Check if a data frame is in wide format
}
\examples{
monthly_data_example |> is_wide()
monthly_data_example |>
  tsbox::ts_long() |>
  is_wide()
dat_in <- monthly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_tslist()
wide_df <- is_wide(dat_in)
x_mod <- if (wide_df) tsbox::ts_long(dat_in) else tsbox::ts_tbl(dat_in)
ans <- if (wide_df) tsbox::ts_wide(x_mod) else tsbox::copy_class(x_mod, dat_in)
}
