% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expr_data.R
\name{load_data_from_feature_counts}
\alias{load_data_from_feature_counts}
\title{load data from featurecounts
borrowed from https://www.biostars.org/p/277316/#277350}
\usage{
load_data_from_feature_counts(filter = TRUE, files = NULL,
  featurecounts_dir, ...)
}
\description{
load data from featurecounts
borrowed from https://www.biostars.org/p/277316/#277350
}
