% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_utils.R
\name{longest_common_substring}
\alias{longest_common_substring}
\title{given a list of strings a, find the longest substring common to all strings}
\usage{
longest_common_substring(a)
}
\arguments{
\item{a}{list or character vector of strings}
}
\description{
given a list of strings a, find the longest substring common to all strings
}
