% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_utils.R
\name{longest_string}
\alias{longest_string}
\title{Return the longest string of a vector/list of strings s}
\usage{
longest_string(s)
}
\arguments{
\item{s}{character or list of strings}
}
\description{
Return the longest string of a vector/list of strings s
}
