% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data.R
\docType{data}
\name{monthly_data_example}
\alias{monthly_data_example}
\title{monthly data for examples}
\format{
tibble of monthly data
}
\source{
author
}
\usage{
monthly_data_example
}
\description{
monthly data for examples
}
\references{
\url{uhero.hawaii.edu}
}
\author{
Peter Fuleky \email{fuleky@hawaii.edu}
}
\keyword{data}
