% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{mtd_cum}
\alias{mtd_cum}
\title{Month to date sum or average}
\usage{
mtd_cum(x, avg = TRUE)
}
\arguments{
\item{x}{a ts-boxable object}

\item{avg}{if TRUE (default), return month to date average, if FALSE, return month to date sum}
}
\value{
object of the same type as the input containing year to date sum or average
}
\description{
Month to date sum or average
}
\examples{
daily_data_example |>
  mtd_cum()
test <- daily_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_pick("VAPNS_HI") |>
  mtd_cum()
tsbox::`\%ts/\%`(test, tsbox::ts_lag(test, "6 months")) |> tail()
}
