% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{mtd_gr}
\alias{mtd_gr}
\title{Month to date growth rate}
\usage{
mtd_gr(long_tbl_in, avg = TRUE)
}
\arguments{
\item{long_tbl_in}{a long tibble of time series (produced by ts_long() for example)}

\item{avg}{if true, the year to date average, if false, the year to date sum for calculation}
}
\value{
a long tibble of time series containing year to date growth rate
}
\description{
Month to date growth rate
}
\examples{
\dontrun{
get_series(c("VISPNS@HI.D", "VAPNS@HI.D")) |>
  tsbox::ts_long() |>
  mtd_gr() |>
  tail()
}
}
