% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{multi_chain}
\alias{multi_chain}
\title{Extend "ts-boxable" objects by chaining (extension of \code{tsbox::ts_chain()})}
\usage{
multi_chain(x, y, ids = NULL)
}
\arguments{
\item{x}{a "ts-boxable" object to be extended}

\item{y}{a "ts-boxable" object containing data to extend x}

\item{ids}{a character vector with series names to be extended (default: NULL,
in which case all series common to x and y are extended)}
}
\value{
object of the same type as the input with extended series
}
\description{
Extend "ts-boxable" objects by chaining (extension of \code{tsbox::ts_chain()})
}
\details{
This function performs a similar operation to \code{tsbox::ts_chain()},
but for multiple series.
}
\examples{
quarterly_data_example |>
  dplyr::filter(time < as.Date("2000-01-01")) |>
  multi_chain(quarterly_data_example, ids = c("E_NF_HI", "ECT_HI"))
quarterly_data_example |>
  dplyr::filter(time < as.Date("2000-01-01")) |>
  multi_chain(quarterly_data_example)
}
