% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{nqtrs}
\alias{nqtrs}
\title{Calculate number of quarters between two dates yyyyQq or yyyy.q}
\usage{
nqtrs(dat1 = "", dat2 = "")
}
\arguments{
\item{dat1}{date of period start (string: yyyyQq or yyyy.q)}

\item{dat2}{date of period end (string: yyyyQq or yyyy.q)}
}
\value{
numeric length of date range in quarters
}
\description{
Calculate number of quarters between two dates yyyyQq or yyyy.q
}
\examples{
nqtrs("2010Q1", "2020Q4")
nqtrs(2010.1, 2020.4)
}
