% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{p}
\alias{p}
\title{Concatenate dates to obtain period}
\usage{
p(dat1 = "", dat2 = "")
}
\arguments{
\item{dat1}{date of period start (string: yyyy-mm-dd)}

\item{dat2}{date of period end (string: yyyy-mm-dd)}
}
\value{
string containing date range
}
\description{
Concatenate dates to obtain period
}
\examples{
p("2010-01-01", "2020-01-01")
p(20100101, 20200101)
p(2010.1, 2020.4)
p("2010Q1", "2020Q4")
p(2010, 2020) # for annual period only
}
