% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{pc_to_pca}
\alias{pc_to_pca}
\title{Convert quarterly growth to annualized growth}
\usage{
pc_to_pca(x, freq = 4)
}
\arguments{
\item{x}{ts-boxable object containing quarterly growth (in percent)}

\item{freq}{numeric frequency of the time series e.g. 4 for quarterly}
}
\value{
object of the same type as the input containing annualized growth (in percent)
}
\description{
Convert quarterly growth to annualized growth
}
\examples{
quarterly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_pc() |>
  pc_to_pca() |>
  tail()
tsbox::ts_c(
  quarterly_data_example |>
    tsbox::ts_long() |>
    tsbox::ts_pc() |>
    pc_to_pca(),
  quarterly_data_example |>
    tsbox::ts_long() |>
    tsbox::ts_pca()
) |>
  dplyr::arrange(id, time) |>
  tsbox::ts_wide()
}
