% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{pca_to_pc}
\alias{pca_to_pc}
\title{Convert annualized growth to quarterly growth}
\usage{
pca_to_pc(ser_in)
}
\arguments{
\item{ser_in}{the series containing annualized growth (in percent)}
}
\value{
series containing quarterly growth (in percent)
}
\description{
Convert annualized growth to quarterly growth
}
\examples{
`ncen@us.sola` <- ts(NA_real_, start = 2016, end = 2021, freq = 1) |> tsbox::ts_xts()
`ncen@us.sola`["2016/2021"] <- c(323127513, 325511184, 327891911, 330268840, 332639102, 334998398)
test1 <- AtoQ(`ncen@us.sola`)
tsbox::ts_c(test1 |> tsbox::ts_pca() |> pca_to_pc(), test1 |> tsbox::ts_pc())
}
