% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{plot_1}
\alias{plot_1}
\title{Interactive plot with level and growth rate}
\usage{
plot_1(
  x,
  rng_start = as.character(Sys.Date() - lubridate::years(10)),
  rng_end = as.character(Sys.Date() + lubridate::years(2)),
  height = 300,
  width = 900,
  yoy_gr = TRUE,
  gr_1 = TRUE
)
}
\arguments{
\item{x}{ts-boxable object to plot (e.g. time series of history, oldsol, sol)}

\item{rng_start}{start of zoom range ("YYYY-MM-DD")}

\item{rng_end}{end of the zoom range ("YYYY-MM-DD")}

\item{height}{height of a single panel (px)}

\item{width}{width of a single panel (px)}

\item{yoy_gr}{year-over-year (default) or annualized growth}

\item{gr_1}{only show growth for the first series (default)}
}
\value{
a dygraph plot
}
\description{
Interactive plot with level and growth rate
}
\examples{
monthly_data_example |>
  plot_1()
quarterly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_pick("E_TU_HI", "ECT_HI", "EMN_HI") |>
  plot_1()
}
