% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{plot_2ax}
\alias{plot_2ax}
\title{Interactive lineplot with two axes}
\usage{
plot_2ax(
  x,
  rng_start = as.character(Sys.Date() - lubridate::years(10)),
  rng_end = as.character(Sys.Date() + lubridate::years(2)),
  height = 300,
  width = 900
)
}
\arguments{
\item{x}{ts-boxable object to plot (e.g. time series of history, oldsol, sol)}

\item{rng_start}{start of zoom range ("YYYY-MM-DD")}

\item{rng_end}{end of the zoom range ("YYYY-MM-DD")}

\item{height}{height of a single panel (px)}

\item{width}{width of a single panel (px)}
}
\value{
a dygraph plot
}
\description{
Interactive lineplot with two axes
}
\examples{
monthly_data_example |>
  plot_2ax()
quarterly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_pick("E_TU_HI", "ECT_HI", "EMN_HI") |>
  plot_2ax()
}
