% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{plot_fc}
\alias{plot_fc}
\title{Interactive plot with level and growth rate for forecast series}
\usage{
plot_fc(
  x,
  rng_start = as.character(Sys.Date() - lubridate::years(10)),
  rng_end = as.character(Sys.Date() + lubridate::years(2)),
  add_table = TRUE,
  table_start = rng_start,
  table_end = rng_end,
  height = 300,
  width = 900,
  yoy_gr = TRUE
)
}
\arguments{
\item{x}{ts-boxable object to plot (min 1, max 3 time series) (e.g. current fcst, old fcst, history)}

\item{rng_start}{start of zoom range ("YYYY-MM-DD")}

\item{rng_end}{end of the zoom range ("YYYY-MM-DD")}

\item{add_table}{should a data table be appended to the plot? (default = TRUE)}

\item{table_start}{start of table range ("YYYY-MM-DD") (all data = NULL, default = rng_start)}

\item{table_end}{end of table range ("YYYY-MM-DD") (all data = NULL, default = rng_end)}

\item{height}{height of a single panel (px)}

\item{width}{width of a single panel (px)}

\item{yoy_gr}{year-over-year (default) or annualized growth}
}
\value{
a dygraph plot
}
\description{
Interactive plot with level and growth rate for forecast series
}
\examples{
monthly_data_example |>
  plot_fc()
quarterly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_pick("E_TU_HI", "ECT_HI", "EMN_HI") |>
  plot_fc()
}
