% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{pm}
\alias{pm}
\title{Concatenate dates formatted as yyyyMm or yyyy.m to obtain period}
\usage{
pm(dat1 = "", dat2 = "")
}
\arguments{
\item{dat1}{date of period start (string: yyyyMm or yyyy.m)}

\item{dat2}{date of period end (string: yyyyMm or yyyy.m)}
}
\value{
string containing date range
}
\description{
Concatenate dates formatted as yyyyMm or yyyy.m to obtain period
}
\examples{
pm("2010M1", "2020M4")
pm(2010.1, 2020.4)
pm(2010.1, )
pm(, 2010.1)
}
