% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{pq}
\alias{pq}
\title{Concatenate dates formatted as yyyyQq or yyyy.q to obtain period}
\usage{
pq(dat1 = "", dat2 = "")
}
\arguments{
\item{dat1}{date of period start (string: yyyyQq or yyyy.q)}

\item{dat2}{date of period end (string: yyyyQq or yyyy.q)}
}
\value{
string containing date range
}
\description{
Concatenate dates formatted as yyyyQq or yyyy.q to obtain period
}
\examples{
pq("2010Q1", "2020Q4")
pq(2010.1, 2020.4)
pq(2010.1, )
pq(, 2010.1)
}
