% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{pq_1}
\alias{pq_1}
\title{Concatenate date formatted as yyyyQq or yyyy.q to obtain single period
DON'T USE!!! USE lubridate::yq() INSTEAD}
\usage{
pq_1(dat1 = "")
}
\arguments{
\item{dat1}{date of period start (string: yyyyQq or yyyy.q)}
}
\value{
string containing date range
}
\description{
Concatenate date formatted as yyyyQq or yyyy.q to obtain single period
DON'T USE!!! USE lubridate::yq() INSTEAD
}
\examples{
pq_1("2010Q1")
pq_1(2010.1)
}
