% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{ptd_cum}
\alias{ptd_cum}
\title{Period to date sum or average}
\usage{
ptd_cum(x, per = "year", avg = TRUE)
}
\arguments{
\item{x}{a ts-boxable object}

\item{per}{unit of time supplied to floor_date() (for ytd per = "year"
(default), for mtd per = "month")}

\item{avg}{if TRUE (default), retorn period to date average, if FALSE, return period to date sum}
}
\value{
object of the same type as the input containing period to date sum or average
}
\description{
Period to date sum or average
}
\examples{
daily_data_example |>
  ptd_cum("week")
test <- daily_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_pick("VAPNS_HI") |>
  ptd_cum("week")
tsbox::`\%ts/\%`(test, tsbox::ts_lag(test, "4 weeks")) |>
  tsbox::`\%ts-\%`(1) |>
  tsbox::`\%ts*\%`(100) |>
  tail()
}
