% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{ptd_gr}
\alias{ptd_gr}
\title{Period to date growth rate}
\usage{
ptd_gr(long_tbl_in, per = "year", lag_length = "1 year", avg = TRUE)
}
\arguments{
\item{long_tbl_in}{a long tibble of time series (produced by ts_long() for example)}

\item{per}{unit of time supplied to floor_date() (for ytd per = "year" (default), for mtd per = "month")}

\item{lag_length}{period over which growth is calculated (e.g. "1 year" (default), "3 years", etc. See ?ts_lag() for options)}

\item{avg}{if true, the year to date average, if false, the year to date sum for calculation}
}
\value{
a long tibble of time series containing year to date growth rate
}
\description{
Period to date growth rate
}
\examples{
\dontrun{
get_series(c("VISNS@HI.M", "VAPNS@HI.M")) |>
  tsbox::ts_long() |>
  ptd_gr() |>
  tail()
get_series("VAPNS@HI.D") |>
  tsbox::ts_long() |>
  ptd_gr(per = "month", lag_length = "3 years") |>
  tail()
}
}
