% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{py}
\alias{py}
\title{Concatenate dates formatted as yyyy to obtain period}
\usage{
py(dat1 = "", dat2 = "")
}
\arguments{
\item{dat1}{year of period start (string or numeric: yyyy)}

\item{dat2}{year of period end (string or numeric: yyyy)}
}
\value{
string containing date range
}
\description{
Concatenate dates formatted as yyyy to obtain period
}
\examples{
py("2010", "2020")
py(2010, 2020)
py(2010, )
py(, 2010)
}
