% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{QtoM}
\alias{QtoM}
\title{Interpolate a tibble of series from quaterly to monthly freq}
\usage{
QtoM(data_q, conv_type)
}
\arguments{
\item{data_q}{tibble containing variables at quarterly freq}

\item{conv_type}{match the quarterly value via "first", "last", "sum", "average"}
}
\value{
tibble containing variables at monthly freq
}
\description{
Interpolate a tibble of series from quaterly to monthly freq
}
\examples{
`ncen@us.sola` <- ts(NA_real_, start = 2016, end = 2021, freq = 1) |> tsbox::ts_xts()
`ncen@us.sola`["2016/2021"] <- c(323127513, 325511184, 327891911, 330268840, 332639102, 334998398)
test1 <- AtoQ(`ncen@us.sola`)
QtoM(tsbox::ts_tbl(test1), "average")
tsbox::ts_frequency(QtoM(tsbox::ts_tbl(test1), "average") |> tsbox::ts_xts())
}
