% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{QtoM_1}
\alias{QtoM_1}
\title{Interpolate a single series from quarterly to monthly freq}
\usage{
QtoM_1(var_q, ts_start, conv_type)
}
\arguments{
\item{var_q}{vector containing a single variable at quarterly freq}

\item{ts_start}{starting period as c(year, quarter) e.g. c(2001, 1)}

\item{conv_type}{match the quarterly value via "first", "last", "sum", "average"}
}
\value{
vector containing a single variable at monthly freq
}
\description{
Interpolate a single series from quarterly to monthly freq
}
\examples{
`ncen@us.sola` <- ts(NA_real_, start = 2016, end = 2021, freq = 1) |> tsbox::ts_xts()
`ncen@us.sola`["2016/2021"] <- c(323127513, 325511184, 327891911, 330268840, 332639102, 334998398)
test1 <- AtoQ(`ncen@us.sola`)
QtoM_1(test1, c(2010, 1), "average")
}
