% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{qtrs}
\alias{qtrs}
\title{Convert period in quarters to period in months}
\usage{
qtrs(nr_quarters)
}
\arguments{
\item{nr_quarters}{number of quarters in period (integer)}
}
\value{
number of months in period
}
\description{
Convert period in quarters to period in months
}
\examples{
qtrs(3)
lubridate::ymd("2020-01-01") + qtrs(3)
}
