% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data.R
\docType{data}
\name{quarterly_data_example}
\alias{quarterly_data_example}
\title{quarterly data for examples}
\format{
tibble of quarterly data
}
\source{
author
}
\usage{
quarterly_data_example
}
\description{
quarterly data for examples
}
\references{
\url{uhero.hawaii.edu}
}
\author{
Peter Fuleky \email{fuleky@hawaii.edu}
}
\keyword{data}
