% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{rename_udaman}
\alias{rename_udaman}
\title{Format series names to udaman format (mnemonic@loc.freq)}
\usage{
rename_udaman(ser_in, freq = NULL)
}
\arguments{
\item{ser_in}{series name (string "mnemonic_loc", "mnemonic__loc_freq")}

\item{freq}{frequency of the series, required if not contained in the series
name (string "D", "W", "M", "Q", "S", "A")}
}
\value{
series name following udaman convention "mnemonic@loc.freq"
}
\description{
Format series names to udaman format (mnemonic@loc.freq)
}
\examples{
rename_udaman(c("E_NF_HI", "ECT_HI", "E_TU_HAW"), freq = "M")
rename_udaman(c("E_NF__HI_M", "ECT__HI_M", "VAP__HAW_W"))
rename_udaman(c("E_NF@HI.M", "ECT@HI.M", "VAP@HAW.W"))
rename_udaman(c("SH_US@HI.M", "SH_JP__HON_M"))
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_series_exp(74) |> dplyr::rename_with(~ rename_udaman(., freq = "M"), .cols = -1)
get_series_exp(74) |> dplyr::rename_with(rename_udaman, freq = "M", .cols = !time)
get_series_exp(74) |>
  tsbox::ts_long() |>
  dplyr::mutate(id = rename_udaman(id, freq = "M"))
get_series_exp(318, rename = "full") |>
  tsbox::ts_long() |>
  dplyr::mutate(id = rename_udaman(id)) |>
  tsbox::ts_xts()
\dontshow{\}) # examplesIf}
}
