% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{set_udaman_token}
\alias{set_udaman_token}
\title{Set udaman token for API access}
\usage{
set_udaman_token(key)
}
\arguments{
\item{key}{a string containing 44 characters}
}
\value{
true if setting the token in .Renviron succeeded
}
\description{
Set udaman token for API access
}
\details{
Save the token in .Renviron as udaman_token = key.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
set_udaman_token("-ABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890=")
\dontshow{\}) # examplesIf}
}
