% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{span}
\alias{span}
\title{Specify span of time series (wrapper around tsbox::ts_span())}
\usage{
span(x, start = NULL, end = NULL, template = NULL, extend = FALSE)
}
\arguments{
\item{x}{ts-boxable object to filter by span}

\item{start}{start date (see examples)}

\item{end}{end date (see examples)}

\item{template}{ts-boxable time series (see tsbox::ts_span)}

\item{extend}{logical. If true, the start and end values are allowed to extend the series (by adding NA values).}
}
\value{
filtered object of the same type as the input
}
\description{
Specify span of time series (wrapper around tsbox::ts_span())
}
\examples{
quarterly_data_example |>
  span(2010.1)
quarterly_data_example |>
  span(2010.1, 2010.4)
quarterly_data_example |>
  span("2010-01-01", "2010-12-31")
}
