% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data.R
\docType{data}
\name{sum_pattern}
\alias{sum_pattern}
\title{end of data range in the data bank}
\format{
string of patterns in series names that should be aggregated via sum
}
\source{
author
}
\usage{
sum_pattern
}
\description{
end of data range in the data bank
}
\references{
\url{uhero.hawaii.edu}
}
\author{
Peter Fuleky \email{fuleky@hawaii.edu}
}
\keyword{data}
