% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data.R
\docType{data}
\name{uh_colors_50}
\alias{uh_colors_50}
\title{transparent versions of UHERO colors}
\format{
vector of hex color codes
}
\source{
author
}
\usage{
uh_colors_50
}
\description{
transparent versions of UHERO colors
}
\references{
\url{uhero.hawaii.edu}
}
\author{
Peter Fuleky \email{fuleky@hawaii.edu}
}
\keyword{data}
