% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data.R
\docType{data}
\name{uhcolors}
\alias{uhcolors}
\title{colors defined in the UHERO Style Guide}
\format{
vector of hex color codes
}
\source{
author
}
\usage{
uhcolors
}
\description{
colors defined in the UHERO Style Guide
}
\references{
\url{uhero.hawaii.edu}
}
\author{
Peter Fuleky \email{fuleky@hawaii.edu}
}
\keyword{data}
