% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data.R
\docType{data}
\name{uhero_colors_light}
\alias{uhero_colors_light}
\title{lighter versions of UHERO colors}
\format{
vector of hex color codes
}
\source{
author
}
\usage{
uhero_colors_light
}
\description{
lighter versions of UHERO colors
}
\references{
\url{uhero.hawaii.edu}
}
\author{
Peter Fuleky \email{fuleky@hawaii.edu}
}
\keyword{data}
