% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{generate volcano plot from deseq2 result}
\usage{
volcano_plot(DESeq2_result, alpha = 0.01, min_fold_change = 2)
}
\description{
generate volcano plot from deseq2 result
}
