% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_filecache.R
\name{with_filecache}
\alias{with_filecache}
\title{Execute an expression, while saving result to a local file}
\usage{
with_filecache(expr, filename, cache_dir = fcutils_options()$cache_dir,
  parse = NULL)
}
\arguments{
\item{expr}{R expression whose result should be cached}

\item{filename}{A unique name for cached item}

\item{cache_dir}{The directory in which to cache result. Defaults to '.Rcache' (to change default, use \code{fcutils_options})}

\item{parse}{(optional) Whether to parse expression before executing (default TRUE for character exprs, otherwise FALSE)}
}
\value{
The contents of the cache-file, if that file exists. Otherwise, result of expr
}
\description{
Execute an expression, while saving result to a local file
}
\examples{
res <- with_filecache(some_long_running_function(),
                     filename = 'cache_this_result.Rds',
                     cache_dir = 'name_of_this_document')
}
