% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapup_for_iSEE.R
\name{wrapup_for_iSEE}
\alias{wrapup_for_iSEE}
\title{Prepare summarized experiment from deseq2 result, for use in iSEE
borrowed from https://gist.github.com/federicomarini/4a543eebc7e7091d9169111f76d59de1}
\usage{
wrapup_for_iSEE(dds, res)
}
\description{
Prepare summarized experiment from deseq2 result, for use in iSEE
borrowed from https://gist.github.com/federicomarini/4a543eebc7e7091d9169111f76d59de1
}
