% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{write_tsd}
\alias{write_tsd}
\title{Save an xts or a wide data frame with time series in tsd format}
\usage{
write_tsd(x, file)
}
\arguments{
\item{x}{a ts-boxable object (only M, Q, A frequency)}

\item{file}{string denoting the location and name of the output file}
}
\value{
nothing (silently save the contents of the tsd file to a user defined
location)
}
\description{
Save an xts or a wide data frame with time series in tsd format
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
quarterly_data_example |> write_tsd("out.tsd")
\dontshow{\}) # examplesIf}
}
