% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{ymd_to_yQq}
\alias{ymd_to_yQq}
\title{Convert dates from yyyy-mm-dd to yyyyQqq format}
\usage{
ymd_to_yQq(x)
}
\arguments{
\item{x}{dates (string: yyyy-mm-dd)}
}
\value{
formatted dates (string: yyyyQqq)
}
\description{
Convert dates from yyyy-mm-dd to yyyyQqq format
}
\examples{
ymd_to_yQq(c("2010-01-01", "2020-10-01"))
ymd_to_yQq(c("2010-01-01", "2020-10-01")) |> lubridate::yq()
}
