% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{ytd_cum}
\alias{ytd_cum}
\title{Year to date sum or average}
\usage{
ytd_cum(x, avg = TRUE)
}
\arguments{
\item{x}{a ts-boxable object}

\item{avg}{if TRUE (default), return year to date average, if FALSE, return year to date sum}
}
\value{
object of the same type as the input containing year to date sum or average
}
\description{
Year to date sum or average
}
\examples{
monthly_data_example |>
  ytd_cum()
monthly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_pick("VISNS_HI") |>
  tsbox::ts_xts() |>
  ytd_cum(avg = FALSE) |>
  tsbox::ts_plot()
}
