% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{ytd_gr}
\alias{ytd_gr}
\title{Year to date growth rate}
\usage{
ytd_gr(x)
}
\arguments{
\item{x}{a ts-boxable object}
}
\value{
object of the same type as the input containing year to date growth rate
}
\description{
Year to date growth rate
}
\examples{
monthly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_pick("VISNS_HI") |>
  tsbox::ts_xts() |>
  ytd_gr() |>
  tail()
}
