% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{AtoQ}
\alias{AtoQ}
\title{Linear interpolation based on aremos command reference page 292}
\usage{
AtoQ(ser_in, aggr = "mean")
}
\arguments{
\item{ser_in}{the xts series to be interpolated (freq = a)}

\item{aggr}{interpolation method: aggregate via mean (default) or sum}
}
\value{
interpolated xts series (freq = q)
}
\description{
Linear interpolation based on aremos command reference page 292
}
\examples{
`ncen@us.sola` <- ts(NA_real_, start = 2016, end = 2021, freq = 1) |> tsbox::ts_xts()
`ncen@us.sola`["2016/2021"] <- c(323127513, 325511184, 327891911, 330268840, 332639102, 334998398)
test1 <- AtoQ(`ncen@us.sola`)
}
