% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{conv_long}
\alias{conv_long}
\title{Convert "ts-boxable" objects into long format}
\usage{
conv_long(x, long_form_only = FALSE)
}
\arguments{
\item{x}{a "tx-boxable" object to be converted}

\item{long_form_only}{should only the long form be returned (TRUE) or additional
details too (default: FALSE)}
}
\value{
a list(long_form, was_wide, ser_names), where
x_mod is a ts-boxable object in long format with id, time and value columns,
was_wide is TRUE if x is a wide data frame, FALSE otherwise,
ser_names are the names of the series in x.
}
\description{
Convert "ts-boxable" objects into long format
}
\details{
This function converts wide data frames and other ts-boxable objects
to the long format (wide data frames are not ts-boxable). In addition, it
ensures that objects containing a single time series have an id column.
}
\examples{
quarterly_data_example |>
  conv_long()
quarterly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_xts() |>
  conv_long()
quarterly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_pick("E_NF_HI") |>
  tsbox::ts_xts() |>
  conv_long()
quarterly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_xts() |>
  tsbox::ts_pick("E_NF_HI") |>
  conv_long()
}
