% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{copy_tbl}
\alias{copy_tbl}
\title{Copy a data frame to clipboard}
\usage{
copy_tbl(x)
}
\arguments{
\item{x}{tibble (or data frame) to be copied}
}
\value{
copy_tbl() returns the input x invisibly
}
\description{
Copy a data frame to clipboard
}
\examples{
monthly_data_example |> copy_tbl()
}
