% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{extract_data}
\alias{extract_data}
\title{Parse gets output and extract underlying data (GETS model development)}
\usage{
extract_data(model_in, y_name)
}
\arguments{
\item{model_in}{a model estimated by arx, isat, or getsm}

\item{y_name}{the actual name of the y variable}
}
\value{
an xts containing the model variables
}
\description{
Parse gets output and extract underlying data (GETS model development)
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# save the data associated with a gets model
\dontshow{\}) # examplesIf}
}
