% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcutils-package.R
\docType{package}
\name{fcutils-package}
\alias{fcutils}
\alias{fcutils-package}
\title{fcutils: Utility Functions for Time Series and Forecasting}
\description{
Manipulate time series. Retrieve and store data. Extend tsbox with additional frequency conversions and growth calculations. Plot time series data. Modify bimets models.
}
\author{
\strong{Maintainer}: Peter Fuleky \email{fuleky@hawaii.edu} (\href{https://orcid.org/0000-0002-4094-2823}{ORCID})

}
\keyword{internal}
