% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcutils-package.R
\docType{package}
\name{fcutils-package}
\alias{fcutils}
\alias{fcutils-package}
\title{fcutils: Time Series and Forecast Utilities}
\description{
Utility functions to retrieve data from UDAMAN, wrangle, transform and plot time series data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://uhero.github.io/fcutils/}
}

}
\author{
\strong{Maintainer}: Peter Fuleky \email{fuleky@hawaii.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
