% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcutils-package.R
\docType{package}
\name{fcutils-package}
\alias{fcutils}
\alias{fcutils-package}
\title{fcutils: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://uhero.github.io/fcutils/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
