% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{find_end}
\alias{find_end}
\title{Find the date of the last observation (NAs are dropped)}
\usage{
find_end(ser_in)
}
\arguments{
\item{ser_in}{an xts series}
}
\value{
date associated with last observation
}
\description{
Find the date of the last observation (NAs are dropped)
}
\examples{
`ncen@us.sola` <- ts(NA_real_, start = 2016, end = 2060, freq = 1) |> tsbox::ts_xts()
`ncen@us.sola`["2016/2018"] <- c(323127513, 325511184, 327891911)
find_end(`ncen@us.sola`)
}
