% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{find_start}
\alias{find_start}
\title{Find the date of the first observation (NAs are dropped)}
\usage{
find_start(x)
}
\arguments{
\item{x}{ts-boxable object}
}
\value{
dates associated with first observation
}
\description{
Find the date of the first observation (NAs are dropped)
}
\examples{
quarterly_data_example |>
  dplyr::mutate(E_NF_HI = dplyr::if_else(time < "2000-01-01", NA_real_, E_NF_HI)) |>
  find_start()
}
