% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{get_series}
\alias{get_series}
\title{Download a set of series from udaman using series names}
\usage{
get_series(
  ser_id_vec,
  format = "wide",
  expand = "true",
  rename = "compact",
  freq = NULL,
  descr = FALSE
)
}
\arguments{
\item{ser_id_vec}{vector of series names}

\item{format}{"wide" (default) or "long" or "xts"}

\item{expand}{"true" (default) or "raw" ("true" downloads formatted data,
"raw" downloads raw units)}

\item{rename}{"compact" (default), "full", "no". "compact": @ replaced by _
and no frequency; "full": @ replaced by __ and . by _; "no": no renaming,
keep UDAMAN names}

\item{freq}{if frequency is missing from series names (or want to modify freq
in existing names) specify frequency, e.g. "M".}

\item{descr}{if TRUE add to the udaman series name the series description in
parentheses}
}
\value{
time and data for all series combined in an object specified by the
format option
}
\description{
Download a set of series from udaman using series names
}
\details{
This function requires access permission to UDAMAN.
Store the udaman token in the .Renviron file using the following format:
udaman_token = "this is your UDAMAN token"
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_series(c("VISNS@HI.M", "VAPNS@HI.M"))
get_series(c("VISNS@HI.M", "VISUSNS@HI.M"), freq = "Q")
get_series(c("VISNS@HI.M", "VAPNS@HI.M"), format = "xts")
get_series(c("VISNS@HI.M"), format = "xts")
get_series(c("VISNS@HI.M"), format = "xts", descr = TRUE)
get_series(c("E_NF_HI", "ECT_HI", "E_TU_HAW"), freq = "M")
get_series(c("E_NF__HI_M", "ECT__HI_M", "VAP__HI_W"))
\dontshow{\}) # examplesIf}
}
