% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{get_series_exp}
\alias{get_series_exp}
\title{Download series listed in an export table from udaman}
\usage{
get_series_exp(
  exp_id,
  format = "wide",
  save_loc = "data/raw",
  expand = "true",
  rename = "compact",
  descr = FALSE
)
}
\arguments{
\item{exp_id}{export id}

\item{format}{"wide" (default) or "long" or "xts"}

\item{save_loc}{location to save the csv of the retrieved data, set to NULL
to avoid saving}

\item{expand}{"true" or "raw" ("true" downloads formatted data, "raw"
downloads raw units)}

\item{rename}{"compact" (default), "full", "no". "compact": @ replaced by _
and no frequency; "full": @ replaced by __ and . by _; "no": no renaming,
keep UDAMAN names}

\item{descr}{if TRUE add to the udaman series name the series description in
parentheses}
}
\value{
time and data for all series combined in a tibble
}
\description{
Download series listed in an export table from udaman
}
\examples{
\dontrun{
get_series_exp(exp_id = 74)
get_series_exp(74, format = "xts")
get_series_exp(74, save_loc = NULL)
}
}
