% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{get_series_exp}
\alias{get_series_exp}
\title{Download series listed in an export table from udaman}
\usage{
get_series_exp(
  exp_id,
  format = "wide",
  expand = "true",
  rename = "compact",
  descr = FALSE,
  save_loc = NULL
)
}
\arguments{
\item{exp_id}{export id (character or numeric)}

\item{format}{"wide" (default) or "long" or "xts"}

\item{expand}{"true" or "raw" ("true" downloads formatted data, "raw" downloads raw units)}

\item{rename}{"compact" (default), "full", "no". "compact": @ replaced by _
and no frequency; "full": @ replaced by \emph{AT} and . by _; "no": no renaming,
keep UDAMAN names}

\item{descr}{if TRUE add to the udaman series name the series description in parentheses (default: FALSE)}

\item{save_loc}{file path for saving data incl. extension ("html" or "csv") (default NULL)}
}
\value{
time and data for all series combined in a tibble
}
\description{
Download series listed in an export table from udaman
}
\details{
This function requires access permission to UDAMAN.
Store the udaman token in the .Renviron file using the following format:
udaman_token = "this is your UDAMAN token"
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_series_exp(exp_id = 74)
get_series_exp(74, format = "xts")
\dontshow{\}) # examplesIf}
}
