% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{ma}
\alias{ma}
\title{Backward looking moving average}
\usage{
ma(x, order)
}
\arguments{
\item{x}{ts-boxable object}

\item{order}{numeric order (window length) of moving average, includes contemporaneous observation}
}
\value{
object of the same type as the input containing moving average
}
\description{
Backward looking moving average
}
\examples{
quarterly_data_example |>
  ma(4) |>
  head()
}
