% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{mtd_cum}
\alias{mtd_cum}
\title{Month to date sum or average}
\usage{
mtd_cum(long_tbl_in, avg = TRUE)
}
\arguments{
\item{long_tbl_in}{a long tibble of time series (produced by ts_long() for
example)}

\item{avg}{if true, the year to date average, if false, the year to date sum}
}
\value{
a long tibble of time series containing year to date sum or average
}
\description{
Month to date sum or average
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_series(c("VISPNS@HI.D", "VAPNS@HI.D")) |>
  tsbox::ts_long() |>
  mtd_cum()
test <- get_series("VAPNS@HI.D") |>
  tsbox::ts_long() |>
  mtd_cum()
test \%ts/\% tsbox::ts_lag(test, "3 years") |> tail()
\dontshow{\}) # examplesIf}
}
