% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{mtd_gr}
\alias{mtd_gr}
\title{Month to date growth rate}
\usage{
mtd_gr(x)
}
\arguments{
\item{x}{a ts-boxable object}
}
\value{
object of the same type as the input containing month to date growth rate
}
\description{
Month to date growth rate
}
\examples{
daily_data_example |>
  mtd_gr() |>
  tail()
}
