% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{pc_to_pca}
\alias{pc_to_pca}
\title{Convert quarterly growth to annualized growth}
\usage{
pc_to_pca(ser_in)
}
\arguments{
\item{ser_in}{series containing quarterly growth (in percent)}
}
\value{
the series containing annualized growth (in percent)
}
\description{
Convert quarterly growth to annualized growth
}
\examples{
`ncen@us.sola` <- ts(NA_real_, start = 2016, end = 2021, freq = 1) |> tsbox::ts_xts()
`ncen@us.sola`["2016/2021"] <- c(323127513, 325511184, 327891911, 330268840, 332639102, 334998398)
test1 <- AtoQ(`ncen@us.sola`)
tsbox::ts_c(test1 |> tsbox::ts_pc() |> pc_to_pca(), test1 |> tsbox::ts_pca())
}
