% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{pca_to_pc}
\alias{pca_to_pc}
\title{Convert annualized growth to quarterly growth}
\usage{
pca_to_pc(x, freq = 4)
}
\arguments{
\item{x}{ts-boxable object containing annualized growth (in percent)}

\item{freq}{numeric frequency of the time series e.g. 4 for quarterly}
}
\value{
object of the same type as the input containing quarterly growth (in percent)
}
\description{
Convert annualized growth to quarterly growth
}
\examples{
quarterly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_pca() |>
  pca_to_pc() |>
  tail()
tsbox::ts_c(
  quarterly_data_example |>
    tsbox::ts_long() |>
    tsbox::ts_pca() |>
    pca_to_pc(),
  quarterly_data_example |>
    tsbox::ts_long() |>
    tsbox::ts_pc()
) |>
  dplyr::arrange(id, time) |>
  tsbox::ts_wide()
}
