% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{pcmp}
\alias{pcmp}
\title{Calculate multi-period average growth}
\usage{
pcmp(x, lag = 4)
}
\arguments{
\item{x}{ts-boxable object for which growth is calculated (in levels)}

\item{lag}{number of period over which growth is calculated}
}
\value{
object of the same type as the input ontaining the average growth of x (in percent)
}
\description{
Calculate multi-period average growth
}
\examples{
quarterly_data_example |>
  pcmp(20) |>
  tail()
}
