% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{plot_2ax}
\alias{plot_2ax}
\title{Interactive lineplot with two axes}
\usage{
plot_2ax(
  ser,
  rng_start = as.character(Sys.Date() - lubridate::years(15)),
  rng_end = as.character(Sys.Date()),
  height = 300,
  width = 900
)
}
\arguments{
\item{ser}{time series to plot (e.g. history, oldsol, sol)}

\item{rng_start}{start of zoom range ("YYYY-MM-DD")}

\item{rng_end}{end of the zoom range ("YYYY-MM-DD")}

\item{height}{height of a single panel (px)}

\item{width}{width of a single panel (px)}
}
\value{
a dygraph plot
}
\description{
Interactive lineplot with two axes
}
\examples{
`ncen@us.sola` <- ts(NA_real_, start = 2016, end = 2021, freq = 1) |> tsbox::ts_xts()
`ncen@us.sola`["2016/2021"] <- c(323127513, 325511184, 327891911, 330268840, 332639102, 334998398)
`ncen@us.oldsola` <- `ncen@us.sola`
`ncen@us.oldsola`["2020/2021"] <- c(352639102, 374998398)
`ncen@us.a` <- `ncen@us.sola`
`ncen@us.a`["2020/2021"] <- NA
test1 <- tsbox::ts_tslist(tsbox::ts_c(`ncen@us.sola`, `ncen@us.oldsola`, `ncen@us.a`)) |>
  purrr::map(AtoQ) |>
  purrr::reduce(tsbox::ts_c) |>
  magrittr::set_names(c("ncen@us.sola", "ncen@us.oldsola", "ncen@us.a"))
plot_2ax(tsbox::ts_c(`ncen@us.sola`, `ncen@us.oldsola`, `ncen@us.a`), rng_start = "2017-01-01")
plot_2ax(test1, rng_start = "2017-01-01")
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_series_exp(74, rename = "no") |>
  tsbox::ts_long() |>
  tsbox::ts_xts() |>
  magrittr::extract(, c("E_NF@HI.Q", "ECT@HI.Q", "EMN@HI.Q")) |>
  plot_2ax()
\dontshow{\}) # examplesIf}
}
