% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{plot_comp_2}
\alias{plot_comp_2}
\title{Two-panel plot of levels, index, and growth rates}
\usage{
plot_comp_2(
  sers,
  rng_start = as.character(Sys.Date() - lubridate::years(15)),
  rng_end = as.character(Sys.Date()),
  height = 300,
  width = 900,
  yoy_gr = TRUE,
  gr_bar = FALSE
)
}
\arguments{
\item{sers}{a vector of series to plot}

\item{rng_start}{start of the zoom range ("YYYY-MM-DD")}

\item{rng_end}{end of the zoom range ("YYYY-MM-DD")}

\item{height}{height of a single panel (px)}

\item{width}{width of a single panel (px)}

\item{yoy_gr}{year-over-year (default) or annualized growth}

\item{gr_bar}{show bars or line (default) for the growth series}
}
\value{
a list with two dygraph plots (level, index, growth)
}
\description{
Two-panel plot of levels, index, and growth rates
}
\examples{
`ncen@us.sola` <- ts(NA_real_, start = 2016, end = 2021, freq = 1) |> tsbox::ts_xts()
`ncen@us.sola`["2016/2021"] <- c(323127513, 325511184, 327891911, 330268840, 332639102, 334998398)
test1 <- AtoQ(`ncen@us.sola`)
plot_comp_2(tsbox::ts_c(`ncen@us.sola`, test1), rng_start = "2017-01-01")
\dontrun{
get_series_exp(74, save_loc = NULL) |>
  tsbox::ts_long() |>
  tsbox::ts_xts() |>
  magrittr::extract(, c("E_NF@HI.Q", "ECT@HI.Q", "EMN@HI.Q")) |>
  plot_comp_2()
}
}
