% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{plot_comp_3}
\alias{plot_comp_3}
\title{Three-panel plot of levels, index, and growth rates}
\usage{
plot_comp_3(
  x,
  indx_start = as.character(Sys.Date() - lubridate::years(10)),
  rng_start = as.character(Sys.Date() - lubridate::years(10)),
  rng_end = as.character(Sys.Date() + lubridate::years(2)),
  height = 300,
  width = 900,
  yoy_gr = TRUE,
  gr_bar = FALSE
)
}
\arguments{
\item{x}{ts-boxable object to plot}

\item{indx_start}{base period for the indexed series ("YYYY-MM-DD")}

\item{rng_start}{start of the zoom range ("YYYY-MM-DD")}

\item{rng_end}{end of the zoom range ("YYYY-MM-DD")}

\item{height}{height of a single panel (px)}

\item{width}{width of a single panel (px)}

\item{yoy_gr}{year-over-year (default) or annualized growth}

\item{gr_bar}{show bars or line (default) for the growth series}
}
\value{
a list with three dygraph plots (level, index, growth)
}
\description{
Three-panel plot of levels, index, and growth rates
}
\examples{
monthly_data_example |>
  plot_comp_3()
quarterly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_pick("E_TU_HI", "ECT_HI", "EMN_HI") |>
  plot_comp_3()
}
