% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{ptd_cum}
\alias{ptd_cum}
\title{Period to date sum or average}
\usage{
ptd_cum(long_tbl_in, per = "year", avg = TRUE)
}
\arguments{
\item{long_tbl_in}{a long tibble of time series (produced by ts_long() for
example)}

\item{per}{unit of time supplied to floor_date() (for ytd per = "year"
(default), for mtd per = "month")}

\item{avg}{if true (default), the year to date average, if false, the year to
date sum}
}
\value{
a long tibble of time series containing year to date sum or average
}
\description{
Period to date sum or average
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_series(c("VISNS@HI.M", "VAPNS@HI.M")) |>
  tsbox::ts_long() |>
  ptd_cum()
\dontshow{\}) # examplesIf}
}
