% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{QtoA}
\alias{QtoA}
\title{Conversion from quarterly to annual frequency}
\usage{
QtoA(ser_in, aggr = "mean")
}
\arguments{
\item{ser_in}{the xts series to be converted (freq = q)}

\item{aggr}{aggregate via mean (default) or sum}
}
\value{
converted xts series (freq = a)
}
\description{
Conversion from quarterly to annual frequency
}
\examples{
`ncen@us.sola` <- ts(NA_real_, start = 2016, end = 2021, freq = 1) |> tsbox::ts_xts()
`ncen@us.sola`["2016/2021"] <- c(323127513, 325511184, 327891911, 330268840, 332639102, 334998398)
test1 <- AtoQ(`ncen@us.sola`)
test2 <- QtoA(test1) # for stock type variables mean, for flow type variables sum
print(test1)
print(cbind(`ncen@us.sola`, test2))
}
