% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{save_plot_list}
\alias{save_plot_list}
\title{Save a list of interactive plots to html}
\usage{
save_plot_list(plot_list, save_loc)
}
\arguments{
\item{plot_list}{a list of plots generated by fcutils::plot_xxxx() functions}

\item{save_loc}{location to save the plots to, including file name}
}
\value{
nothing (silently save the html to a user defined location)
}
\description{
Save a list of interactive plots to html
}
\examples{
# hold the plots in a list
plot_out <- list()
for (i in monthly_data_example[2:3] |> names()) {
  plot_out[[i]] <- plot_1(monthly_data_example |> tsbox::ts_long() |>
                             dplyr::filter(stringr::str_detect(id, i)),
                             rng_start = as.character(Sys.Date() - lubridate::years( 5)),
                             rng_end = as.character(Sys.Date() + lubridate::years(7)),
                             width = 1500, height = 650, yoy_gr = TRUE)
}
# specify location of the output
save_loc <- stringr::str_c("~/Downloads/plots_", Sys.Date(), ".html")
# combine a list of plots into a single html
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
plot_out \%>\% save_plot_list(save_loc)
\dontshow{\}) # examplesIf}
}
