% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data.R
\docType{data}
\name{sum_pattern}
\alias{sum_pattern}
\title{pattern in series names that should be aggregate as sum}
\format{
string of patterns in series names that should be aggregated via sum
}
\source{
author
}
\usage{
sum_pattern
}
\description{
pattern in series names that should be aggregate as sum
}
\references{
\url{uhero.hawaii.edu}
}
\author{
Peter Fuleky \email{fuleky@hawaii.edu}
}
\keyword{data}
