% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{write_tsd}
\alias{write_tsd}
\title{Save an xts or a wide data frame with time series in tsd format}
\usage{
write_tsd(x, file)
}
\arguments{
\item{x}{an xts or a (wide) data frame with time series (only M, Q, A
frequency)}

\item{file}{string denoting the location and name of the output file}
}
\value{
nothing (silently save the contents of the tsd file to a user defined
location)
}
\description{
Save an xts or a wide data frame with time series in tsd format
}
\examples{
\dontrun{
get_series_exp(74) |> write_tsd("out.tsd")
}
}
