% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst_utils.R
\name{yoy_to_lev}
\alias{yoy_to_lev}
\title{Extend a series using year over year growth}
\usage{
yoy_to_lev(yoy_gr, hist_lev)
}
\arguments{
\item{yoy_gr}{ts-boxable object containing year over year growth rates}

\item{hist_lev}{ts-boxable object containing the history in levels
for forecast and at least one year of history (in percent)}
}
\value{
object of the same type as hist_lev extended with year over year growth
}
\description{
Extend a series using year over year growth
}
\details{
This function only works for univariate time series and requires
that the growth rates are available for at least the last year of history.
}
\examples{
quarterly_data_example |>
  tsbox::ts_long() |>
  dplyr::filter(id == "E_NF_HI") |>
  tsbox::ts_pcy() |>
  yoy_to_lev(
    quarterly_data_example |>
      tsbox::ts_long() |>
      dplyr::filter(id == "ECT_HI", time <= "2010-01-01")
  )
}
